move the btree package
discussed with Gilles and Mathieu
