+++ snapshot/errors.go
@@ -3,9 +3,9 @@ package snapshot
 import (
 	"strings"
 
+	"github.com/PlakarKorp/plakar/btree"
 	"github.com/PlakarKorp/plakar/objects"
 	"github.com/PlakarKorp/plakar/packfile"
-	"github.com/PlakarKorp/plakar/snapshot/btree"
 	"github.com/vmihailenco/msgpack/v5"
 )
